Armor: Moved details to detail package, refactored set loading to not require deterministic loading

+++ app/src/main/java/com/ghstudios/android/data/database/DataManager.java
@@ -173,7 +173,7 @@ public class DataManager {
         return mHelper.queryComponentCreated(id);
     }
 
-    public ComponentCursor queryComponentCreateByArmorFamily(int familyId){
+    public ComponentCursor queryComponentCreateByArmorFamily(long familyId){
         return itemDao.queryComponentsByArmorFamily(familyId);
     }
